RB_Technology/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.3.0
requests>=2.26.0
flask>=2.0.0
rembg>=2.0.0
torchgeometry>=0.1.2
scipy>=1.7.0
